Rewrite InternalCheckbox and Checkbox
