+++ packages/bento-design-system/src/CheckboxField/createCheckboxField.tsx
@@ -4,7 +4,7 @@ import { useRef } from "react";
 import { FieldType } from "../Field/createField";
 import { Children } from "..";
 import { FieldProps } from "../Field/FieldProps";
-import { createInternalCheckbox } from "../Checkbox/createInternalCheckbox";
+import { InternalCheckbox } from "../Checkbox/InternalCheckbox";
 import { useField } from "@react-aria/label";
 import { SelectionControlConfig } from "../Field/Config";
 
@@ -16,8 +16,6 @@ export function createCheckboxField(
   config: SelectionControlConfig,
   { Field }: { Field: FieldType }
 ) {
-  const Checkbox = createInternalCheckbox(config);
-
   return function CheckboxField(props: Props) {
     const checkboxProps = {
       ...props,
@@ -42,7 +40,7 @@ export function createCheckboxField(
         assistiveTextProps={{}}
         errorMessageProps={errorMessageProps}
       >
-        <Checkbox option={checkboxProps} inputProps={inputProps} inputRef={ref} />
+        <InternalCheckbox option={checkboxProps} inputProps={inputProps} inputRef={ref} />
       </Field>
     );
   };
