+++ packages/bento-design-system/src/CheckboxGroupField/createCheckboxGroupField.tsx
@@ -2,7 +2,7 @@ import { useCheckboxGroup, useCheckboxGroupItem } from "@react-aria/checkbox";
 import { useRef } from "react";
 import { FieldType } from "../Field/createField";
 import { FieldProps } from "../Field/FieldProps";
-import { createInternalCheckbox } from "../Checkbox/createInternalCheckbox";
+import { createInternalCheckbox } from "../Checkbox/InternalCheckbox";
 import { useField } from "@react-aria/label";
 import { CheckboxGroupState, useCheckboxGroupState } from "@react-stately/checkbox";
 import { Box, Inline, Stack } from "../internal";
